cleancode修改
# message auto-generated for no-merge-commit merge:
merge ******_cleancode into master

cleancode修改

Created-by: insightDeveloper

Match-id-b91f2b14c7c6cfe0fca275aaecdab71709448da5

+++ server/src/modules/summary/core/ExpertHotspotManager.h
@@ -19,7 +19,7 @@ struct FillExpertDataParams {
     std::vector<ExpertDeploymentStruct> deployment;
     std::vector<int> moeLayerMapping;
     int colNumber = 0;
-    int expertNumberPerRank = 0;
+    uint64_t expertNumberPerRank = 0;
     std::set<int> denseLayerSet;
 };
